Restructure completions
