+++ vhdl_lang/src/completion/tokenizer.rs
@@ -1,0 +1,111 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+// Copyright (c) 2024, Olof Kraigher olof.kraigher@gmail.com
+use crate::data::ContentReader;
+use crate::syntax::{Symbols, Tokenizer};
+use crate::{Position, Source, Token};
+
+/// Tokenizes `source` up to `cursor` but no further. The last token returned is the token
+/// where the cursor currently resides or the token right before the cursor.
+///
+/// Examples:
+///
+/// input = "use ieee.std_logic_1|164.a"
+///                              ^ cursor position
+/// `tokenize_input(input)` -> {USE, ieee, DOT, std_logic_1164}
+///
+/// input = "use ieee.std_logic_1164|.a"
+///                                 ^ cursor position
+/// `tokenize_input(input)` -> {USE, ieee, DOT, std_logic_1164}
+///
+/// input = "use ieee.std_logic_1164.|a"
+///                                  ^ cursor position
+/// `tokenize_input(input)` -> {USE, ieee, DOT, std_logic_1164, DOT}
+/// input = "use ieee.std_logic_1164.a|"
+///                                   ^ cursor position
+/// `tokenize_input(input)` -> {USE, ieee, DOT, std_logic_1164, DOT, a}
+///
+/// On error, or if the source is empty, returns an empty vector.
+pub(crate) fn tokenize_input(symbols: &Symbols, source: &Source, cursor: Position) -> Vec<Token> {
+    let contents = source.contents();
+    let mut tokenizer = Tokenizer::new(symbols, source, ContentReader::new(&contents));
+    let mut tokens = Vec::new();
+    loop {
+        match tokenizer.pop() {
+            Ok(Some(token)) => {
+                if token.pos.start() >= cursor {
+                    break;
+                }
+                tokens.push(token);
+            }
+            Ok(None) => break,
+            Err(_) => return vec![],
+        }
+    }
+    tokens
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::syntax::test::Code;
+    use crate::syntax::Kind::*;
+    use assert_matches::assert_matches;
+
+    macro_rules! kind {
+        ($kind: pat) => {
+            crate::syntax::Token { kind: $kind, .. }
+        };
+    }
+
+    #[test]
+    fn tokenizing_an_empty_input() {
+        let input = Code::new("");
+        let tokens = tokenize_input(&input.symbols, input.source(), Position::new(0, 0));
+        assert_eq!(tokens.len(), 0);
+    }
+
+    #[test]
+    fn tokenizing_stops_at_the_cursors_position() {
+        let input = Code::new("use ieee.std_logic_1164.all");
+        let mut cursor = input.s1("std_logic_11").pos().end();
+        let tokens = tokenize_input(&input.symbols, input.source(), cursor);
+        assert_matches!(
+            tokens[..],
+            [kind!(Use), kind!(Identifier), kind!(Dot), kind!(Identifier)]
+        );
+        cursor = input.s1("std_logic_1164").pos().end();
+        let tokens = tokenize_input(&input.symbols, input.source(), cursor);
+        assert_matches!(
+            tokens[..],
+            [kind!(Use), kind!(Identifier), kind!(Dot), kind!(Identifier)]
+        );
+        cursor = input.s1("std_logic_1164.").pos().end();
+        let tokens = tokenize_input(&input.symbols, input.source(), cursor);
+        assert_matches!(
+            tokens[..],
+            [
+                kind!(Use),
+                kind!(Identifier),
+                kind!(Dot),
+                kind!(Identifier),
+                kind!(Dot)
+            ]
+        );
+        cursor = input.s1("std_logic_1164.all").pos().end();
+        let tokens = tokenize_input(&input.symbols, input.source(), cursor);
+        assert_matches!(
+            tokens[..],
+            [
+                kind!(Use),
+                kind!(Identifier),
+                kind!(Dot),
+                kind!(Identifier),
+                kind!(Dot),
+                kind!(All)
+            ]
+        );
+    }
+}
